Consolidate request-input bounds tests into one file
Merge the stop-string, stop-token-id, bad-words, beam-search, and
DeepSeek history-scan regression tests into a single
tests/test_request_input_bounds.py per review.
